requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
ultralytics==8.0.200
hydra-core>=1.2.0
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
chardet==5.2.0

easyocr

# Logging -------------------------------------
tensorboard>=2.4.1


# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation

# HUB -----------------------------------------
GitPython>=3.1.24